# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LLMProxy
  version: 0.2.0
  description: OpenAI-compatible inference gateway with capability routing, local operator roles, accounting, reports and durable batches. See api.md and expanded-api.md for compatibility boundaries.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:4000
security:
  - gatewayKey: []
paths:
  /v1/models:
    get:
      operationId: listModels
      summary: List permitted and configured public model aliases
      responses:
        '200':
          description: Model list
          content:
            application/json:
              schema:
                type: object
                required: [object, data]
                properties:
                  object: { type: string, const: list }
                  data:
                    type: array
                    items:
                      type: object
                      required: [id, object, created, owned_by]
                      properties:
                        id: { type: string }
                        object: { type: string, const: model }
                        created: { type: integer }
                        owned_by: { type: string, const: llmproxy }
        '401': { $ref: '#/components/responses/OpenAiError' }
        '429': { $ref: '#/components/responses/OpenAiError' }
        '503': { $ref: '#/components/responses/OpenAiError' }
...
  /admin/auth/login:
    post:
      operationId: loginOperator
      summary: Create an eight-hour local operator session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorLogin'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorSession'
        default:
          $ref: '#/components/responses/Problem'
  /admin/auth/me:
    get:
      operationId: currentOperator
      summary: Read the current administrative identity
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      - auditor
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  username:
                    type: string
                  role:
                    type: string
                    enum:
                    - administrator
                    - operator
                    - auditor
        default:
          $ref: '#/components/responses/Problem'
  /admin/auth/logout:
    post:
      operationId: logoutOperator
      summary: Revoke the current local session
      description: Shared bootstrap secrets are not revoked by logout.
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      - auditor
      responses:
        '204':
          description: Successful response
        default:
          $ref: '#/components/responses/Problem'
  /admin/operators:
    get:
      operationId: listOperators
      summary: List local operators without password hashes
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Operator'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createOperator
      summary: Create a local operator
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOperator'
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operator'
        default:
          $ref: '#/components/responses/Problem'
  /admin/operators/{id}:
    put:
      operationId: updateOperator
      summary: Update a local operator and revoke existing sessions
      description: The last enabled local administrator cannot be disabled or demoted.
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOperator'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operator'
        default:
          $ref: '#/components/responses/Problem'
  /admin/keys/{id}/rotate:
    post:
      operationId: rotateApiKey
      summary: Rotate a credential while preserving its key ID and counters
      description: Only the current and immediately previous credentials can remain valid. The new key is disclosed
        once.
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                grace_seconds:
                  type: integer
                  minimum: 0
                  maximum: 86400
                  default: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedApiKey'
        default:
          $ref: '#/components/responses/Problem'
  /admin/keys/page:
    get:
      operationId: pageApiKeys
      summary: Page key metadata by exact owner
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      - auditor
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - name: owner
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyPage'
        default:
          $ref: '#/components/responses/Problem'
  /admin/keys/{id}/quotas:
    get:
      operationId: listQuotaWindows
      summary: Read up to 120 UTC monthly quota windows
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      - auditor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuotaWindow'
        default:
          $ref: '#/components/responses/Problem'
  /admin/usage/page:
    get:
      operationId: pageUsage
      summary: Page retained usage with stable timestamp and ID ordering
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      - auditor
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - name: api_key_id
        in: query
        schema:
          type: string
          format: uuid
      - name: owner
        in: query
        schema:
          type: string
      - name: from
        in: query
        schema:
          type: string
          format: date-time
        description: Inclusive timestamp
      - name: to
        in: query
        schema:
          type: string
          format: date-time
        description: Exclusive timestamp
      - name: model
        in: query
        schema:
          type: string
      - name: provider
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
          enum:
          - success
          - error
          - cancelled
          - abandoned
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsagePage'
        default:
          $ref: '#/components/responses/Problem'
  /admin/usage/summary:
    get:
      operationId: summarizeUsage
      summary: Aggregate all matching retained usage by model and provider
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      - auditor
      parameters:
      - name: api_key_id
        in: query
        schema:
          type: string
          format: uuid
      - name: owner
        in: query
        schema:
          type: string
      - name: from
        in: query
        schema:
          type: string
          format: date-time
        description: Inclusive timestamp
      - name: to
        in: query
        schema:
          type: string
          format: date-time
        description: Exclusive timestamp
      - name: model
        in: query
        schema:
          type: string
      - name: provider
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
          enum:
          - success
          - error
          - cancelled
          - abandoned
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsageRollup'
        default:
          $ref: '#/components/responses/Problem'
  /admin/usage/export:
    get:
      operationId: exportUsage
      summary: Stream matching usage as CSV across all pages
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      - auditor
      parameters:
      - name: api_key_id
        in: query
        schema:
          type: string
          format: uuid
      - name: owner
        in: query
        schema:
          type: string
      - name: from
        in: query
        schema:
          type: string
          format: date-time
        description: Inclusive timestamp
      - name: to
        in: query
        schema:
          type: string
          format: date-time
        description: Exclusive timestamp
      - name: model
        in: query
        schema:
          type: string
      - name: provider
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
          enum:
          - success
          - error
          - cancelled
          - abandoned
      responses:
        '200':
          description: Successful response
          content:
            text/csv:
              schema:
                type: string
        default:
          $ref: '#/components/responses/Problem'
  /admin/usage/{id}/attempts:
    get:
      operationId: listUpstreamAttempts
      summary: List recorded HTTP attempts for a request
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      - auditor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UpstreamAttempt'
        default:
          $ref: '#/components/responses/Problem'
  /admin/audit:
    get:
      operationId: pageAudit
      summary: Page management intent/outcome and rejected-request audit records
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      - auditor
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditPage'
        default:
          $ref: '#/components/responses/Problem'
  /admin/billing/reconcile:
    post:
      operationId: reconcileInvoice
      summary: Atomically reconcile provider attempt costs
      description: Identical imports are idempotent. Conflicting attempt IDs or references return 409. Adjusts request
        and lifetime/original-month costs; retained usage is required.
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ReconciliationInput'
              minItems: 1
              maxItems: 1000
      responses:
        '204':
          description: Successful response
        default:
          $ref: '#/components/responses/Problem'
  /admin/config/reload:
    post:
      operationId: reloadConfiguration
      summary: Validate and replace model, pricing and account catalogs
      description: Invalid configuration returns 400 and preserves the active catalog. Applies per instance. Process
        settings require restart; in-flight routes retain captured providers and prices.
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  generation:
                    type: integer
                  models:
                    type: integer
                  providers:
                    type: integer
        default:
          $ref: '#/components/responses/Problem'
  /admin/models:
    get:
      operationId: listManagedModels
      summary: List model routing and effective target feature flags
      security:
      - adminKey: []
      x-required-roles:
      - administrator
      - operator
      - auditor
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    routing:
                      type: string
                    max_concurrent_requests:
                      type: integer
                    targets:
                      type: array
                      items:
                        type: object
                        properties:
                          provider:
                            type: string
                          model:
                            type: string
                          capabilities:
                            type: string
        default:
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    gatewayKey:
      type: http
      scheme: bearer
      bearerFormat: Gateway API key
    adminKey:
      type: http
      scheme: bearer
      bearerFormat: Local operator session or separate LLMPROXY_ADMIN_KEY bootstrap secret
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor from the previous page; URL-encode without modification
  responses:
    OpenAiError:
      description: Sanitized OpenAI-compatible error
      headers:
        X-Request-Id:
          schema: { type: string }
        Retry-After:
          description: Seconds until admission may be available; sent for gateway RPM and concurrency denials.
          schema: { type: integer }
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [message, type, param, code]
                properties:
                  message: { type: string }
                  type: { type: string }
                  param: { type: [string, 'null'] }
                  code: { type: string }
    Problem:
      description: RFC 7807 problem details
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type: { type: string }
              title: { type: string }
              status: { type: integer }
              code: { type: string }
              request_id: { type: string }
...
  schemas:
...
    KeyPolicy:
      type: object
      required: [allowed_models]
      properties:
        allowed_models:
          type: array
          minItems: 1
          items: { type: string }
        requests_per_minute: { type: integer, minimum: 1, maximum: 1000000, default: 60 }
        token_limit: { type: [integer, 'null'], minimum: 0 }
        spending_budget: { type: [number, 'null'], minimum: 0, maximum: 1000000 }
        expires_at: { type: [string, 'null'], format: date-time }
        monthly_token_limit: { type: [integer, 'null'], minimum: 0 }
        monthly_spending_budget: { type: [number, 'null'], minimum: 0, maximum: 1000000 }
    ApiKeySummary:
      allOf:
        - $ref: '#/components/schemas/KeyPolicy'
        - type: object
          properties:
            id: { type: string, format: uuid }
            prefix: { type: string }
            owner: { type: string }
            enabled: { type: boolean }
            used_tokens: { type: integer }
            reserved_tokens: { type: integer }
            spent: { type: number }
            reserved_cost: { type: number }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            last_used_at: { type: [string, 'null'], format: date-time }
    UsageRecord:
      type: object
      properties:
        operation: { type: string, enum: [chat, embeddings, responses] }
        request_id: { type: string, format: uuid }
        api_key_id: { type: string, format: uuid }
        model: { type: string }
        provider: { type: string }
        input_tokens: { type: integer }
        output_tokens: { type: integer }
        total_tokens: { type: integer }
        latency_ms: { type: integer }
        status: { type: string, enum: [success, error, cancelled, abandoned] }
        error_code: { type: [string, 'null'] }
        estimated_cost: { type: number }
        usage_estimated: { type: boolean }
        created_at: { type: string, format: date-time }
...
    Operator:
      type: object
      required:
      - id
      - username
      - role
      - enabled
      - created_at
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum:
          - administrator
          - operator
          - auditor
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateOperator:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[A-Za-z0-9._@-]+$
        password:
          type: string
          minLength: 12
          maxLength: 1024
          writeOnly: true
        role:
          type: string
          enum:
          - administrator
          - operator
          - auditor
          default: operator
    UpdateOperator:
      type: object
      required:
      - enabled
      - role
      properties:
        enabled:
          type: boolean
        role:
          type: string
          enum:
          - administrator
          - operator
          - auditor
        password:
          type:
          - string
          - 'null'
          minLength: 12
          maxLength: 1024
          writeOnly: true
    OperatorLogin:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
        password:
          type: string
          writeOnly: true
    OperatorSession:
      type: object
      required:
      - token
      - expires_at
      - operator
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        operator:
          $ref: '#/components/schemas/Operator'
    CreatedApiKey:
      type: object
      required:
      - key
      - details
      properties:
        key:
          type: string
          pattern: ^llmp_sk_[A-Za-z0-9_-]+$
        details:
          $ref: '#/components/schemas/ApiKeySummary'
    QuotaWindow:
      type: object
      properties:
        id:
          type: string
        api_key_id:
          type: string
          format: uuid
        period:
          type: string
          pattern: ^\d{4}-\d{2}$
        used_tokens:
          type: integer
        reserved_tokens:
          type: integer
        spent_units:
          type: integer
          description: Nano-USD; 1,000,000,000 units equal USD 1
        reserved_units:
          type: integer
          description: Nano-USD
    UsageRollup:
      type: object
      properties:
        model:
          type: string
        provider:
          type: string
        requests:
          type: integer
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        cost:
          type: number
          minimum: 0
        average_latency_ms:
          type: number
    AuditRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        actor:
          type: string
        action:
          type: string
        resource:
          type: string
        status_code:
          type: integer
          minimum: 0
          maximum: 599
          description: Zero for intent records; HTTP status for outcome records
        request_id:
          type:
          - string
          - 'null'
          format: uuid
    UpstreamAttempt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        request_id:
          type: string
          format: uuid
        provider:
          type: string
        model:
          type: string
        provider_key_id:
          type:
          - string
          - 'null'
          pattern: '^key_[a-f0-9]{32}$'
          description: Stable upstream credential fingerprint (first 128 bits of SHA-256); null for historical, identity/keyless or non-HTTP attempts. Never contains the raw API key.
        provider_request_id:
          type:
          - string
          - 'null'
        created_at:
          type: string
          format: date-time
        latency_ms:
          type: integer
        status:
          type: string
        http_status:
          type:
          - integer
          - 'null'
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        cached_input_tokens:
          type: integer
        cache_creation_tokens:
          type: integer
        usage_estimated:
          type: boolean
        estimated_cost:
          type: number
          minimum: 0
        actual_cost:
          type:
          - number
          - 'null'
          minimum: 0
    ReconciliationInput:
      type: object
      required:
      - attempt_id
      - actual_cost
      - reference
      properties:
        attempt_id:
          type: string
          format: uuid
        actual_cost:
          type: number
          minimum: 0
          maximum: 1000000
        reference:
          type: string
          minLength: 1
          maxLength: 128
    UsagePage:
      type: object
      required:
      - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UsageRecord'
        next_cursor:
          type:
          - string
          - 'null'
      description: Pass next_cursor unchanged as the URL-encoded cursor parameter. Its absence or null ends pagination.
    KeyPage:
      type: object
      required:
      - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeySummary'
        next_cursor:
          type:
          - string
          - 'null'
      description: Pass next_cursor unchanged as the URL-encoded cursor parameter. Its absence or null ends pagination.
    AuditPage:
      type: object
      required:
      - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditRecord'
        next_cursor:
          type:
          - string
          - 'null'
      description: Pass next_cursor unchanged as the URL-encoded cursor parameter. Its absence or null ends pagination.